removed unused element, reorganized helper functions to respective files
